requirements.txt
django==1.9
gunicorn==19.4.5
django-taggit
djangorestframework